[REFACTOR] Move shell_templates from ui to core package
