+++ .agents/tasks/taskmaster/[CODER] Audit and reduce env var config usage.md
