+++ agents_runner/core/shell_templates.py
