+++ agents_runner/docker/agent_worker.py
@@ -39,8 +39,8 @@ from agents_runner.docker.env_image_builder import ensure_env_image
 from agents_runner.prompts import load_prompt
 from agents_runner.log_format import format_log
 from agents_runner.log_format import wrap_container_log
-from agents_runner.ui.shell_templates import git_identity_clause
-from agents_runner.ui.shell_templates import shell_log_statement
+from agents_runner.core.shell_templates import git_identity_clause
+from agents_runner.core.shell_templates import shell_log_statement
 from agents_runner.midoriai_template import MidoriAITemplateDetection
 from agents_runner.midoriai_template import scan_midoriai_agents_template
