+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "midori-ai-agents-runner"
-version = "0.1.0.6"
+version = "0.1.0.7"
 description = "A GUI for running AI agents in Docker containers with workspace and GitHub management."
 readme = "README.md"
 requires-python = "==3.13.11"
